services/video_analysis/tasks.py
import time
import os
import tempfile
import asyncio
import whisper
from celery_app import celery_app
from common.logger import get_logger

logger = get_logger("video_analysis_tasks")

# Load Whisper model once at worker startup (using the "base" model as an example)
model = whisper.load_model("base")

def transcribe_audio_sync(audio_path: str) -> str:
    """
    Synchronously transcribe audio using the loaded Whisper model.
    """
    try:
        result = model.transcribe(audio_path)
        transcription = result.get("text", "")
        return transcription
    except Exception as e:
        logger.error(f"Transcription failed: {e}")
        return ""

def diarization_processing(file_path: str) -> dict:
    """
    Synchronous placeholder for speaker diarization.
    Replace with an integrated diarization library (e.g., pyannote.audio).
    """
    time.sleep(0.5)
    return {"speaker_1": "Speaker 1", "speaker_2": "Speaker 2"}

def perform_diarization_sync(audio_path: str) -> dict:
    """
    Synchronously perform diarization using a placeholder function.
    """
    try:
        return diarization_processing(audio_path)
    except Exception as e:
        logger.error(f"Diarization failed: {e}")
        return {}

def extract_insights_sync(transcription: str, diarization: dict) -> list:
    """
    Synchronously run NLP processing to extract insights.
    Replace with real NLP pipelines (e.g., SpaCy or HuggingFace Transformers).
    """
    time.sleep(0.5)
    return [{
        "timestamp": time.time(),
        "speaker": diarization.get("speaker_1", "Unknown"),
        "action": "Follow up on discussion",
        "keywords": ["follow", "discussion"],
        "sentiment": "positive"
    }]

@celery_app.task
def process_audio_chunk(audio_bytes: bytes):
    """
    Celery task to process an audio chunk.
    Writes the bytes to a temp file, then synchronously runs transcription,
    diarization, and insight extraction.
    """
    with tempfile.NamedTemporaryFile(suffix=".wav", delete=False) as tmp:
        tmp.write(audio_bytes)
        tmp.flush()
        tmp_name = tmp.name

    try:
        transcription = transcribe_audio_sync(tmp_name)
        diarization = perform_diarization_sync(tmp_name)
        insights = extract_insights_sync(transcription, diarization)
        return insights
    except Exception as e:
        logger.error(f"Error processing audio chunk: {e}")
        return []
    finally:
        os.unlink(tmp_name) 