services/video_analysis/requirements.txt
fastapi
uvicorn
pydantic
openai-whisper 